// src/models/ClusterResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ClusterOut } from './ClusterOut';
export type ClusterResponse = {
    /**
     * Идентификатор запроса
     */
    response_id?: string;
    /**
     * Кластер
     */
    cluster: ClusterOut;
};