// src/models/NodeGroupResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { NodeGroupOut } from './NodeGroupOut';
export type NodeGroupResponse = {
    /**
     * Идентификатор запроса
     */
    response_id?: string;
    /**
     * Группа нод
     */
    node_group: NodeGroupOut;
};