// src/models/ResourcesResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Resources } from './Resources';
export type ResourcesResponse = {
    /**
     * Идентификатор запроса
     */
    response_id?: string;
    /**
     * Ресурсы кластера
     */
    resources: Resources;
};